Started adding the shared interface
